requirements.txt
azure.identity
azure.monitor.query
azure.mgmt.rdbms
azure.mgmt.sql
azure.mgmt.storage
pandas
openpyxl